Teach androidtestrunner running android bundles
Add the modern android bundles support to androidtestrunner.
The tool uses Android bundle tool to unpack and install Android
bundles. If the Android bundletool found in the system and
QT_USE_ANDROID_MODERN_BUNDLE is enabled Qt tests prefer run 'aab'
but not APKs.

This unlocks testing of Android bundles in CI and locally.

Change-Id: I98be662f5c383127a8401371757ea708a495a94f
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>