Fix flaky tst_QFocusEvent::checkReason_ActiveWindow on openSUSE
tst_QFocusEvent::CheckReason_ActiveWindow is flaky on openSUSE where
the childFocusWidgetOne is unable to receive focus after d has been
hidden.

This is due to some asynchronicity when setActive is being called
while d is already active.

Remove the call to activateWindow on d as it is already active by
calling show.

Pick-to: 6.9 6.8 6.5
Change-Id: Id01c6704122df42982d65d28ddb94d184d00d9aa
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>