Http2: check if stream was RSTed w/ bespoke function
We have a function to check if a stream was locally reset - we should
realistically only process HEADERS on RSTed streams for a limited amount
of time. I forgot we had this mechanism in place already, let's update
to use it for the recent fix.

Amends 9224fdd3f4cccce7a8123ab398f6ff8745d759ce

Pick-to: 6.9.1 6.9 6.8
Change-Id: I4826100704542f681341ceb84c4f85332f8ae03d
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>