Introduce _qt_android_in_finalizer property
The property indicates that CMake target currently evaluates the
target Android finalizer. This property is useful to detect potential
misuse of the properties that expected to be set at any time before
finalizing the Android target but used only when finalizing the target.

Property designed to store the finalizer type for further extention
capabilities.

Change-Id: I75239358c566a7aa9ae33423474210aaf54d3020
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>