CMake: Fix rpath adjustment when deploying QML plugins
If Qt is configured with a custom value for INSTALL_QMLDIR, e.g. via
"configure -qmldir lib/qml", then we must adjust the rpath for QML
plugins. Otherwise they won't be able to load Qt libraries.

We had general code that adjusts the rpath for plugins, but that only
triggered if INSTALL_LIBDIR was different from the user project's
CMAKE_INSTALL_LIBDIR.

Now, we always adjust the rpath of plugins on Linux. The operation is
cheap, and I don't quite see how to add a robust condition for all
desired cases.

Pick-to: 6.8 6.9
Fixes: QTBUG-125223
Change-Id: Ib511fbc2b733deac804191773e56be9b6adf5e68
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>