QString: replace(pos, len ~~~) should have a precondition len >= 0
It makes no sense to have a negative length here, it most likely is a
result of a calculation error in user code.

Fix the issue by documenting the precondition and enforcing it in debug
builds with Q_PRE.

Remove the test data rows that were supposed to be fixed; now we assert
that `len` isn't negative.

QByteArray should behave the same, but due to historic/backwards
compatibility issues it treats a negative `len` as if it's `0`, i.e. it
inserts `after` at pos.

Pick-to: 6.10
Change-Id: I46ad5934412aeed86a94a3753c42bda262af1cf9
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>