QPainter: make Private::d_ptrs a QVLA<unique_ptr>
This way, we remove the impedance mismatch between QPainter::d_ptr
(which already is a unique_ptr) and Private::d_ptrs, which still dealt
in raw pointers, simplifying the code.

Since we're now owning the objects in d_ptrs, add an assertion that
d_ptrs is empty on QPainter destruction.

Coverity-Id: 425479
Change-Id: I849e80752b4bc5a71aef8d47663c6ffc9fdca9f7
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>