Add function to generate AndroidManifest from template
This now generates AndroidManifest.xml for the Qt applications in CMake
space if QT_USE_ANDROID_MODERN_BUNLDE is enabled. The generated CMake
AndroidManifest.xml has less functionality comparing to the one that is
generated by androiddeployqt. But the missing parts will be implemented
later.

Known issue:
androiddeployqt still provides its own AndroidManifest.xml, so there is
an issue in the iterative builds, since there is no declaration in CMake
that androiddeployqt actually writes this file. This will be fixed in
further patchsets.

Change-Id: I219b7f7c85a76c070b0c98c6fe72a68a38fe92bf
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>