wayland: Deliver WindowScreenChangedEvent synchronously
The QWaylandWindow::handleScreensChanged() function checks
window()->screen(), but QWindow's screen is updated asynchronously. It
can lead to the following issue:

- start on screen 1
- QWaylandWindow::handleWindowScreenChanged(screen 2): it'll update
  the window geometry so the window position is at the top left corner
  of screen 2
- QWaylandWindow::handleWindowScreenChanged(screen 1): it'll do nothing,
  despite the fact that the geometry has to be updated

This change makes the delivery of WindowScreenChanged events sync, so
the QPA updates the window geometry as expected if the window moves back
and forth between two outputs.

This fixes misplaced popup menus in Konsole.

Change-Id: I5a200b33ef63c6082f54ee85667b7b1ab513335e
Reviewed-by: David Edmundson <davidedmundson@kde.org>