Q(ObjectBindable)Property: Properly support comparisons
Before, it was possible to do some comparisons due to the implicit
conversion to T. However, that does not work in all contexts.

Fix this by explicitly declaring the comparison operators. We need to
a) use the helper macro directly, as the operartors are only available
   if the underlying type has them and
b) mark the operator as noexcept(false), as capturing a property in a
   binding might allocate.

Fixes: QTBUG-134921
Change-Id: I2855964ba481b9e7778a4a7076528593549910fe
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>