Release AndroidManifest.xml template from genex placeholders
Read the template and substitute symbols that are sensitive for
generator expression processor with the respective generator
expressions implictily, instead of forcing users to use them in
templates.

Side effect of this change - users cannot specify generator expressions
directly in templates anymore, not that we already anounced this
feature, but is worth mentioning.

Change-Id: I565b0c0780211801467627ed4f3de3629b8efa08
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>