Fix flaky tst_QScroller::overshoot on multiple Windows versions
tst_QScroller::overshoot is flaky on multiple versions of Windows 11.
The failure occurs in the part testing overshoot always off and slow,
where the result of the deltaPos and the startPos are equal to 0
instead of being smaller.

Slightly increase the value of touchEnd so that deltaPos is smaller.
As a drive-by, constify a local variable.

Fix the flakiness on Windows 11 24h2 mingw.

Fixes: QTBUG-137201
Change-Id: I6dacc7b5ac140f427fa6fbb8fddb5447ae14a233
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>