wayland: Drop ShellIntegration::handleExpose
This was effectively deprecated when the virtual isExposed was added and
we moved towards a state based system.

The region of the exposed geometry is something we want to phase out and
is of no use to shells.

The other usage in XdgShell was for updating contentGeometry on first
show, but this is now an explicit call updated by QWaylandWindow
directly.

Change-Id: I09a083fcfcc69c444ca4270f425b591b046e0617
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>