QReadWriteLock: inline the locking content
This is similar to QMutex, which has an inline uncontended portion and
leaves the handling of contention (or recursive QRWL) to the out-of-line
content.

Using tst_bench_qreadwritelock on an Intel i9-7940X with the CPUs fixed
to 3.1 GHz and only using the first 4 cores, I noticed:
* The cost for an uncontended lock of either type decreased by 26
  instructions but only 1 cycle.
* Compared to an uncontended QMutex, either type of uncontended lock
  only requires 2 more instructions (11 vs 9).
* Contended read locks appear to improve by about 274 cycles and 35
  instructions, or over 6.7% and 16.5% respectively (excluding the
  "nothing" base cost).
* Contended write locks appear to be slower by about 118 cycles and 8
  instructions, though the benchmark varies a lot.
* Likewise, recursive QReadWriteLocks cost 5 instructions more.

This means that for QReadWriteLocks used where reads are far more
frequent than writes, this is a meaningful performance improvement, at
the cost of some more code being emitted on use. This makes everyone
have roughly the same benefit as QtCore in LTO mode.

The extra cost for contended write locks and for recursion are probably
caused by the checks upon entry to the contendedTryLock* functions,
which duplicate what was inlined.

Change-Id: Ia78b7edb86b31435b7b2fffd82cd5e27688355e8
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>