QRhiPassResourceTracker: store as a QFlapMap instead of QHash
QVarLengthFlatMap is more efficient in this hot path, since it does
not allocate and has contiguous memory.

Now that QRhiPassResourceTracker is a big struct, we use emplace_back()
instead of append(QRhiPassResourceTracker()) to avoid a copy when
creating a new tracker, and store the list as a QVector instead of QVLA.

As a drive-by, add a couple of Q_UNLIKELY to put the warning paths in
cold section.

Change-Id: Ib7df6d6ef618d138774d60a359a482cd9aeb34d9
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>