QObject: (dis)connect(): normalize/decode the same signature
The SIGNAL/SLOT() macros add an integer to the beginning of a
signal/method's signature as a marker; when calling
QMOP::decodeMethodSiganture() we pass the function's signature without
the marker, so logically the same should be done when calling
QMO::normalizedSignature(). The behavior doesn't change, but we don't
inc/decrement the signal/slot pointers back and forth, which makes the
code more readable.

Pick-to: 6.10
Change-Id: I967a78a5a2076ab2dabcbed589b061746cdd39f7
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>